统计：会员统计

+++ yudao-module-mall/yudao-module-statistics-biz/src/main/java/cn/iocoder/yudao/module/statistics/service/trade/bo/TradeOrderSummaryRespBO.java
@@ -13,7 +13,7 @@ public class TradeOrderSummaryRespBO {
     /**
      * 创建订单数
      */
-    private Long orderCreateCount;
+    private Integer orderCreateCount;
     /**
      * 支付订单商品数
      */
